+++ src/main/java/community/flock/dialogflow/ci/dialogflow/helper/RequestHelper.java
@@ -57,8 +57,6 @@ public class RequestHelper {
 		
 		String json = "{}";
 		
-		System.out.println(target.getUri().toString());
-		System.out.println(json);
 		return doRequest(target, json);
 	}
 
@@ -69,8 +67,6 @@ public class RequestHelper {
 				String.format("/projects/%s/agent:import", projectId));
 		
 		String json = generateImportRequestBody(payload);
-		System.out.println(target.getUri().toString());
-		System.out.println(json);
 		
 		doRequest(target, json);
 	}
@@ -85,8 +81,6 @@ public class RequestHelper {
 				      .post(Entity.entity(json, MediaType.APPLICATION_JSON_TYPE), 
 				    		  	String.class);
 
-			System.out.println("RESPONSE:");
-			System.out.println(bean);
 			return bean;
 		} catch(javax.ws.rs.WebApplicationException e) {
 			System.out.println(e.getResponse().readEntity(String.class));